split hybrid_offset.cpp
