+++ inst/include/dplyr/HybridHandlerMap.h
@@ -11,6 +11,7 @@ void install_minmax_handlers(HybridHandlerMap& handlers);
 void install_count_handlers(HybridHandlerMap& handlers);
 void install_nth_handlers(HybridHandlerMap& handlers);
 void install_window_handlers(HybridHandlerMap& handlers);
+void install_offset_handlers(HybridHandlerMap& handlers);
 
 bool hybridable(RObject arg);
